requirements.txt
# Core ML frameworks
torch>=1.12.0
torchvision>=0.13.0
torchaudio>=0.12.0

# CompressAI for neural compression
compressai>=1.2.0

# torchdistill for knowledge distillation
git+https://github.com/yoshitomo-matsubara/torchdistill.git

# Data and image processing
numpy>=1.21.0
pillow>=8.3.0
opencv-python>=4.5.0

# Scientific computing
scipy>=1.7.0
scikit-learn>=1.0.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.1.0
hydra-core>=1.1.0

# Logging and visualization
tensorboard>=2.8.0
wandb>=0.12.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.64.0
rich>=12.0.0

# Development tools
jupyter>=1.0.0
ipython>=7.30.0
black>=22.0.0
flake8>=4.0.0
pytest>=7.0.0

# Optional: for faster data loading
timm>=0.6.0  # For additional model architectures if needed
